mv some bf16 funcs to float16_bfloat16_utils.h
